Improve TTS test fixtures
